Convert chan1m32isis for gtest
