+++ isis/src/chandrayaan1/apps/chan1m32isis/chan1m32isis.h
@@ -1,0 +1,11 @@
+#ifndef chan1m32isis_h
+#define chan1m32isis_h
+
+#include "UserInterface.h"
+#include "Pvl.h"
+
+namespace Isis {
+  extern void chan1m32isis(UserInterface &ui);
+}
+
+#endif
