+++ isis/tests/data/chan1m32isis/forwardAscending/M3T20090630T083407_V03_TIM_cropped.TAB
@@ -1,0 +1,5 @@
+     1 2009-06-30T08:34:35.653371 2009 180.357357101487
+     2 2009-06-30T08:34:35.602491 2009 180.357356512598
+     3 2009-06-30T08:34:35.551611 2009 180.357355923710
+     4 2009-06-30T08:34:35.500731 2009 180.357355334822
+     5 2009-06-30T08:34:35.449851 2009 180.357354745933
