+++ isis/tests/data/chan1m32isis/forwardDescending/M3G20081129T171431_V03_TIM_cropped.TAB
@@ -1,0 +1,5 @@
+     1 2008-11-29T17:14:29.780687 2008 333.718400229561
+     2 2008-11-29T17:14:29.882447 2008 333.718401407338
+     3 2008-11-29T17:14:29.984207 2008 333.718402585115
+     4 2008-11-29T17:14:30.085967 2008 333.718403762892
+     5 2008-11-29T17:14:30.187727 2008 333.718404940670
