+++ isis/tests/data/chan1m32isis/l0/M3G20090106T113423_V01_L0_cropped.LBL
@@ -1,0 +1,69 @@
+PDS_VERSION_ID = PDS3
+LABEL_REVISION_NOTE = "2010-02-09, S. Lundeen"
+DATA_SET_ID = CH1-ORB-L-M3-2-L0-RAW-V1.0
+PRODUCT_ID = M3G20090106T113423_V01_L0
+RECORD_TYPE = UNDEFINED
+MISSION_ID = CH1
+MISSION_NAME = CHANDRAYAAN-1
+INSTRUMENT_HOST_ID = CH1-ORB
+INSTRUMENT_HOST_NAME = "CHANDRAYAAN-1 ORBITER"
+INSTRUMENT_NAME = "MOON MINERALOGY MAPPER"
+INSTRUMENT_ID = M3
+TARGET_NAME = MOON
+TARGET_TYPE = SATELLITE
+MISSION_PHASE_NAME = "PRIMARY MISSION"
+PRODUCT_TYPE = RAW_IMAGE
+PRODUCT_CREATION_TIME = 2009-06-10T05:03:42
+START_TIME = 2009-01-06T11:34:23
+STOP_TIME = 2009-01-06T11:34:47
+SPACECRAFT_CLOCK_START_COUNT = 4/1165041.799
+SPACECRAFT_CLOCK_STOP_COUNT = 4/1165065
+ORBIT_NUMBER = 710
+PRODUCT_VERSION_TYPE = PRELIMINARY
+PRODUCER_INSTITUTION_NAME = "JET PROPULSION LABORATORY"
+SOFTWARE_NAME = m3_igds_l0_v18.pl
+SOFTWARE_VERSION_ID = "18"
+DESCRIPTION = "M3 Level 0 data product which consists of raw science data, reassembled into time-sequenced data in units of digital numbers."
+INSTRUMENT_MODE_ID = GLOBAL
+DETECTOR_TEMPERATURE = 161.26
+CH1:SWATH_WIDTH = 320 <PIXELS>
+CH1:SWATH_LENGTH = 229 <LINES>
+Object = L0_FILE
+  ^L0_LINE_PREFIX_TABLE = M3G20090106T113423_V01_L0.IMG
+  ^L0_IMAGE = M3G20090106T113423_V01_L0_cropped.IMG
+  RECORD_TYPE = FIXED_LENGTH
+  RECORD_BYTES = 1920
+  FILE_RECORDS = 5
+  Object = L0_LINE_PREFIX_TABLE
+    INTERCHANGE_FORMAT = BINARY
+    ROWS = 229
+    COLUMNS = 9
+    ROW_BYTES = 1280
+    ROW_SUFFIX_BYTES = 55040
+    ^STRUCTURE = LN_PRFX_HDR.FMT
+  End_Object
+  Object = L0_IMAGE
+    LINES = 5
+    LINE_SAMPLES = 320
+    LINE_PREFIX_BYTES = 1280
+    SAMPLE_TYPE = LSB_INTEGER
+    SAMPLE_BITS = 16
+    UNIT = DN
+    BANDS = 3
+    BAND_STORAGE_TYPE = LINE_INTERLEAVED
+    LINE_DISPLAY_DIRECTION = DOWN
+    SAMPLE_DISPLAY_DIRECTION = RIGHT
+  End_Object
+End_Object
+Object = L0_HDR_FILE
+  ^L0_ENVI_HEADER = M3G20090106T113423_V01_L0.HDR
+  RECORD_TYPE = VARIABLE_LENGTH
+  FILE_RECORDS = 11
+  Object = L0_ENVI_HEADER
+    INTERCHANGE_FORMAT = ASCII
+    BYTES = 304
+    HEADER_TYPE = ENVI
+    DESCRIPTION = "Header file for compatibility with the commercial software package ENVI."
+  End_Object
+End_Object
+End
