+++ isis/tests/data/chan1m32isis/linerateNotConstant/M3G20081118T223204_V03_TIM_cropped.TAB
@@ -1,0 +1,5 @@
+     1 2008-11-18T22:32:03.461991 2008 322.938928945601
+     2 2008-11-18T22:32:03.563749 2008 322.938930123362
+     3 2008-11-18T22:32:03.665508 2008 322.938931301123
+     4 2008-11-18T22:32:03.767267 2008 322.938932478885
+     5 2008-11-18T22:32:03.869025 2008 322.938933656646
