+++ isis/tests/data/chan1m32isis/reverseAscending/M3G20090423T191900_V03_L1B_cropped.LBL
@@ -1,0 +1,179 @@
+PDS_VERSION_ID = PDS3
+LABEL_REVISION_NOTE = "2009-01-26, S. Lundeen, 2010-12-07, S. Lundeen, 2011-09-13, S. Lundeen"
+DATA_SET_ID = CH1-ORB-L-M3-4-L1B-RADIANCE-V3.0
+PRODUCT_ID = M3G20090423T191900_V03_RDN
+RECORD_TYPE = UNDEFINED
+MISSION_ID = CH1
+MISSION_NAME = CHANDRAYAAN-1
+INSTRUMENT_HOST_ID = CH1-ORB
+INSTRUMENT_HOST_NAME = "CHANDRAYAAN-1 ORBITER"
+INSTRUMENT_NAME = "MOON MINERALOGY MAPPER"
+INSTRUMENT_ID = M3
+TARGET_NAME = MOON
+TARGET_TYPE = SATELLITE
+MISSION_PHASE_NAME = "PRIMARY MISSION"
+PRODUCT_TYPE = CALIBRATED_IMAGE
+PRODUCT_CREATION_TIME = 2011-09-01T02:15:01
+START_TIME = 2009-04-23T19:19:00
+STOP_TIME = 2009-04-23T19:19:43
+SPACECRAFT_CLOCK_START_COUNT = 9/1365722.595
+SPACECRAFT_CLOCK_STOP_COUNT = 9/1365765.843
+ORBIT_NUMBER = 2013
+PRODUCT_VERSION_TYPE = ACTUAL
+PRODUCT_VERSION_ID = "3.0"
+SOURCE_PRODUCT_ID = M3G20090423T191900_V01_L0.IMG
+PRODUCER_INSTITUTION_NAME = "JET PROPULSION LABORATORY"
+SOFTWARE_NAME = m3g_l1b_v04.exe
+SOFTWARE_VERSION_ID = "04"
+DESCRIPTION = "M3 Level 1B data product which contains pixel located, radiometrically-calibrated data."
+SOLAR_DISTANCE = 1.003288249239 <AU>
+INSTRUMENT_MODE_ID = GLOBAL
+DETECTOR_TEMPERATURE = 151.21
+CH1:SWATH_WIDTH = 304 <pixel>
+CH1:SWATH_LENGTH = 426 <pixel>
+CH1:SPACECRAFT_YAW_DIRECTION = REVERSE
+CH1:ORBIT_LIMB_DIRECTION = ASCENDING
+SPACECRAFT_ORIENTATION = (0.203593734991, 0.687904954046, 181.026875788702)
+CH1:INITIAL_SC_ORIENTATION = (N/A, N/A, N/A)
+CH1:SC_ORIENTATION_EPOCH_TDB_TIME = N/A
+CH1:SC_ORIENTATION_RATE = (N/A, N/A, N/A)
+CH1:SC_ROTATION_AXIS_VECTOR = (N/A, N/A, N/A)
+CH1:SC_ROTATION_RATE = N/A
+^DESCRIPTION = L1B_NAV_DESC.ASC
+CH1:SPECTRAL_CALIBRATION_FILE_NAME = M3G20081211_RDN_SPC.TAB
+CH1:RAD_GAIN_FACTOR_FILE_NAME = M3G20081211_RDN_GAIN.TAB
+CH1:GLOBAL_BANDPASS_FILE_NAME = M3G20081211_RDN_BPF.IMG
+Object = RDN_FILE
+  ^RDN_IMAGE = M3G20090423T191900_V03_RDN_cropped.IMG
+  RECORD_TYPE = FIXED_LENGTH
+  RECORD_BYTES = 3648
+  FILE_RECORDS = 5
+  Object = RDN_IMAGE
+    LINES = 5
+    LINE_SAMPLES = 304
+    SAMPLE_TYPE = PC_REAL
+    SAMPLE_BITS = 32
+    UNIT = "W/(m^2 um sr)"
+    BANDS = 3
+    BAND_STORAGE_TYPE = LINE_INTERLEAVED
+    LINE_DISPLAY_DIRECTION = DOWN
+    SAMPLE_DISPLAY_DIRECTION = RIGHT
+  End_Object
+End_Object
+Object = RDN_HDR_FILE
+  ^RDN_ENVI_HEADER = M3G20090423T191900_V03_RDN.HDR
+  RECORD_TYPE = VARIABLE_LENGTH
+  FILE_RECORDS = 1146
+  Object = RDN_ENVI_HEADER
+    INTERCHANGE_FORMAT = ASCII
+    BYTES = 35657
+    HEADER_TYPE = ENVI
+    DESCRIPTION = "Header file for compatibility with the commercial software package ENVI."
+  End_Object
+End_Object
+Object = LOC_FILE
+  ^LOC_IMAGE = M3G20090423T191900_V03_LOC_cropped.IMG
+  RECORD_TYPE = FIXED_LENGTH
+  RECORD_BYTES = 7296
+  FILE_RECORDS = 5
+  Object = LOC_IMAGE
+    LINES = 5
+    LINE_SAMPLES = 304
+    SAMPLE_TYPE = PC_REAL
+    SAMPLE_BITS = 64
+    BANDS = 3
+    BAND_STORAGE_TYPE = LINE_INTERLEAVED
+    BAND_NAME = (Longitude, Latitude, Radius)
+    LINE_DISPLAY_DIRECTION = DOWN
+    SAMPLE_DISPLAY_DIRECTION = RIGHT
+  End_Object
+End_Object
+Object = LOC_HDR_FILE
+  ^LOC_ENVI_HEADER = M3G20090423T191900_V03_LOC.HDR
+  RECORD_TYPE = VARIABLE_LENGTH
+  FILE_RECORDS = 16
+  Object = LOC_ENVI_HEADER
+    INTERCHANGE_FORMAT = ASCII
+    BYTES = 371
+    HEADER_TYPE = ENVI
+    DESCRIPTION = "Header file for compatibility with the commercial software package ENVI."
+  End_Object
+End_Object
+Object = OBS_FILE
+  ^OBS_IMAGE = M3G20090423T191900_V03_OBS_cropped.IMG
+  RECORD_TYPE = FIXED_LENGTH
+  RECORD_BYTES = 12160
+  FILE_RECORDS = 5
+  Object = OBS_IMAGE
+    LINES = 5
+    LINE_SAMPLES = 304
+    SAMPLE_TYPE = PC_REAL
+    SAMPLE_BITS = 32
+    BANDS = 10
+    BAND_STORAGE_TYPE = LINE_INTERLEAVED
+    BAND_NAME = ("To-Sun AZM", "To-Sun Zenith", "To-Inst AZM", "To-Inst Zenith", Phase-angle, "To-Sun Path Length", "To-Inst Path Length", "Facet Slope", "Facet Aspect", "Facet Cos i")
+    LINE_DISPLAY_DIRECTION = DOWN
+    SAMPLE_DISPLAY_DIRECTION = RIGHT
+  End_Object
+End_Object
+Object = OBS_HDR_FILE
+  ^OBS_ENVI_HEADER = M3G20090423T191900_V03_OBS.HDR
+  RECORD_TYPE = VARIABLE_LENGTH
+  FILE_RECORDS = 21
+  Object = OBS_ENVI_HEADER
+    INTERCHANGE_FORMAT = ASCII
+    BYTES = 706
+    HEADER_TYPE = ENVI
+    DESCRIPTION = "Header file for compatibility with the commercial software package ENVI."
+  End_Object
+End_Object
+Object = UTC_FILE
+  ^UTC_TIME_TABLE = M3G20090423T191900_V03_TIM_cropped.TAB
+  RECORD_TYPE = FIXED_LENGTH
+  RECORD_BYTES = 57
+  FILE_RECORDS = 5
+  Object = UTC_TIME_TABLE
+    NAME = "UTC OBSERVATION TIMING DATA"
+    INTERCHANGE_FORMAT = ASCII
+    ROWS = 5
+    COLUMNS = 4
+    ROW_BYTES = 57
+    Object = COLUMN
+      COLUMN_NUMBER = 1
+      NAME = "LINE NUMBER"
+      DATA_TYPE = ASCII_INTEGER
+      START_BYTE = 1
+      BYTES = 6
+      FORMAT = I6
+      DESCRIPTION = "Record number for each RDN image line"
+    End_Object
+    Object = COLUMN
+      COLUMN_NUMBER = 2
+      NAME = UTC_TIME
+      DATA_TYPE = TIME
+      START_BYTE = 8
+      BYTES = 26
+      FORMAT = A26
+      DESCRIPTION = "UTC Time for the middle of the integration period for each RDN image line expressed as YYYY-MM-DDTHH:MM:SS.SSSSSS"
+    End_Object
+    Object = COLUMN
+      COLUMN_NUMBER = 3
+      NAME = YEAR
+      DATA_TYPE = CHARACTER
+      START_BYTE = 35
+      BYTES = 4
+      FORMAT = I4
+      DESCRIPTION = "Decimal Day of Year (DDOY) Year reference extracted from the earliest time of each RDN image line"
+    End_Object
+    Object = COLUMN
+      COLUMN_NUMBER = 4
+      NAME = DDOY
+      DATA_TYPE = DATE
+      START_BYTE = 40
+      BYTES = 16
+      FORMAT = F16.12
+      DESCRIPTION = "Decimal Day of Year represented as the number of days elapsed since 00:00 UTC of January 1 of the year associated with the time stamp of the first line of the RDN image file. DDOY is expressed using seventeen characters where 1-3 = three characters that contain the integer number of days; 4 = a decimal point; 5-16 = twelve charact- ers after the decimal for the fractional part of the day of year value."
+    End_Object
+  End_Object
+End_Object
+End
