+++ isis/tests/data/chan1m32isis/reverseAscending/M3G20090423T191900_V03_TIM_cropped.TAB
@@ -1,0 +1,5 @@
+     1 2009-04-23T19:19:45.137902 2009 112.805383559362
+     2 2009-04-23T19:19:45.036142 2009 112.805382381587
+     3 2009-04-23T19:19:44.934382 2009 112.805381203811
+     4 2009-04-23T19:19:44.832622 2009 112.805380026034
+     5 2009-04-23T19:19:44.730862 2009 112.805378848258
