+++ isis/tests/data/chan1m32isis/reverseDescending/M3G20090106T113423_V03_TIM_cropped.TAB
@@ -1,0 +1,5 @@
+     1 2009-01-06T11:34:24.431536 2009   5.482227218696
+     2 2009-01-06T11:34:24.533296 2009   5.482228396473
+     3 2009-01-06T11:34:24.635056 2009   5.482229574250
+     4 2009-01-06T11:34:24.736816 2009   5.482230752026
+     5 2009-01-06T11:34:24.838576 2009   5.482231929803
